+++ src/main/java/com/education/qporter/global/config/WebConfig.java
@@ -1,0 +1,18 @@
+package com.education.qporter.global.config;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.servlet.config.annotation.CorsRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+@Configuration
+public class WebConfig implements WebMvcConfigurer {
+
+	@Override
+	public void addCorsMappings(CorsRegistry registry) {
+		registry.addMapping("/**") // 모든 API 경로에 대해
+			.allowedOrigins("*") // 모든 origin 허용
+			.allowedMethods("*") // 모든 HTTP 메서드 허용
+			.allowedHeaders("*") // 모든 헤더 허용
+			.allowCredentials(false); // 인증정보 포함 안함
+	}
+}
